rename base auth file index.ts
