+++ src/app/api/auth/[...all]/route.ts
@@ -1,4 +1,4 @@
-import { auth } from '@/lib/auth/auth';
+import { auth } from '@/lib/auth';
 
 const handler = auth.handler;
