+++ src/lib/auth/getUserId.ts
@@ -1,4 +1,4 @@
-import { auth } from '@/lib/auth/auth';
+import { auth } from '@/lib/auth';
 import { headers } from 'next/headers';
 
 export async function getUserId() {
